RavenDB-24702 Add warning when the doc has been updated/deleted
